migrated StateComponent to Beta::State
