+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -90,7 +90,7 @@
   - title: Spinner
     url: "/components/beta/spinner"
   - title: State
-    url: "/components/state"
+    url: "/components/beta/state"
   - title: Subhead
     url: "/components/subhead"
   - title: TabContainer
