+++ static/statuses.json
@@ -62,6 +62,7 @@
   "Primer::Beta::ProgressBar": "beta",
   "Primer::Beta::RelativeTime": "beta",
   "Primer::Beta::Spinner": "beta",
+  "Primer::Beta::State": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
@@ -89,7 +90,7 @@
   "Primer::OcticonSymbolsComponent": "deprecated",
   "Primer::PopoverComponent": "deprecated",
   "Primer::SpinnerComponent": "deprecated",
-  "Primer::StateComponent": "beta",
+  "Primer::StateComponent": "deprecated",
   "Primer::SubheadComponent": "beta",
   "Primer::TabContainerComponent": "alpha",
   "Primer::TimeAgoComponent": "deprecated",
